+++ src/Azure.Functions.Cli/Kubernetes/KEDA/Models/ScaledObjectScaleTargetRefV1Alpha1.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Azure.Functions.Cli.Kubernetes.KEDA
+{
+    public class ScaledObjectScaleTargetRefV1Alpha1
+    {
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("envSourceContainerName")]
+        public string EnvSourceContainerName { get; set; }
+    }
+}
