+++ src/Azure.Functions.Cli/Kubernetes/Models/Kubernetes/IKubernetesSpec.cs
@@ -1,5 +1,6 @@
 namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
 {
     public interface IKubernetesSpec
-    { }
+    { 
+    }
 }
